assignment_01/pyproject.toml
[project]
name = "assignment-01"
version = "0.1.0"
description = "assignment 1: singular value decomposition of images"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "image>=1.5.33",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "pyyaml>=6.0.2",
]
